% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\docType{methods}
\name{calld,RedPort-method}
\alias{calld,RedPort-method}
\alias{calld}
\title{Call RedeR app from R}
\usage{
\S4method{calld}{RedPort}(obj, filepath = "default", maxlag = 20, checkcalls = FALSE)
}
\arguments{
\item{obj}{A \code{RedPort}-class object.}

\item{filepath}{A path to the 'reder.jar' file available in the RedeR
R package <string>}

\item{maxlag}{Max acceptable lag time for the R-Java callback confirmation
(default=20 s) <numeric>.}

\item{checkcalls}{A logical value, whether to report errors from the
R-to-Java calls.}
}
\value{
System call to start the RedeR application.
}
\description{
Method to launch RedeR application from R.
}
\examples{

rdp <- RedPort()

\donttest{
calld(rdp)
}

}
\seealso{
\code{\link{startRedeR}}
}
\author{
Sysbiolab.
}
