% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{Deprecated}
\alias{Deprecated}
\alias{gtoy.rm}
\alias{cea}
\alias{deleteSelectedEdges}
\alias{att.mape}
\alias{deleteSelectedNodes}
\alias{getContainerComponets}
\alias{updateContainerSize}
\alias{updateGraph}
\alias{updateCoordXY}
\alias{deSelectGraph}
\alias{deSelectNodes}
\alias{deSelectEdges}
\alias{selectGraph}
\alias{getNodes}
\alias{getEdges}
\alias{mergeNodes}
\alias{setArrowDirection}
\alias{addEdgeBetweenContainers}
\alias{nesthc}
\alias{getSourceEdgeIDs}
\alias{getTargetEdgeIDs}
\alias{getEdgeIDs}
\alias{selectAllNodes}
\alias{selectAllEdges}
\alias{getNodeIDs}
\alias{addLegend.color}
\alias{addLegend.shape}
\alias{addLegend.size}
\title{Deprecated functions}
\usage{
gtoy.rm(x, ...)

deleteSelectedEdges()

att.mape()

deleteSelectedNodes()

getContainerComponets(x, ...)

updateContainerSize(x, ...)

updateGraph(x, ...)

updateCoordXY(x, ...)

deSelectGraph(x, ...)

deSelectNodes(x, ...)

deSelectEdges(x, ...)

selectGraph(x, ...)

getNodes(x, ...)

getEdges(x, ...)

cea(x, ...)

mergeNodes(x, ...)

setArrowDirection(x, ...)

addEdgeBetweenContainers(x, ...)

nesthc(x, ...)

getSourceEdgeIDs(x, ...)

getTargetEdgeIDs(x, ...)

getEdgeIDs(x, ...)

selectAllNodes(x, ...)

selectAllEdges(x, ...)

getNodeIDs(x, ...)

addLegend.color(x, ...)

addLegend.shape(x, ...)

addLegend.size(x, ...)
}
\arguments{
\item{x}{Deprecated arg.}

\item{...}{Additional deprecated args.}
}
\value{
---
}
\description{
List of deprecated functions in RedeR V3.
}
\examples{
# List of deprecated functions in RedeR V3:
# -gtoy.rm (no replacement)
# -cea (no replacement)
# -mergeNodes (set in app only)
# -setArrowDirection (set via addGraph function)
# -addEdgeBetweenContainers (set in app only)
# -nesthc (no replacement)
# -getSourceEdgeIDs (replaced by getGraph)
# -getTargetEdgeIDs (replaced by getGraph)
# -getEdgeIDs (replaced by getGraph)
# -selectAllNodes (replaced by selectNodes)
# -selectAllEdges (replaced by selectEdges)
# -getNodeIDs (replaced by getGraph)
# -addLegend.color (replaced by addLegend)
# -addLegend.shape (replaced by addLegend)
# -addLegend.size (replaced by addLegend)
# -getContainerComponets (replaced by getGraph)
# -updateContainerSize (set in app only)
# -deleteSelectedEdges (replaced by deleteEdges)
# -deleteSelectedNodes (replaced by deleteNodes)

}
\seealso{
\code{\link{addGraph}}, \code{\link{getGraph}}
}
