\name{duplicateGraph}
\alias{duplicateGraph}
\alias{duplicateGraph,RedPort-method}

\title{Duplicate graphs in RedeR application.}

\description{
Deprecated function...
}

\usage{
duplicateGraph(obj, ...)
}

\arguments{
  \item{obj}{ Object of RedPort Class. }
  \item{...}{ Additional arguments passed to RedeR application. }    
}

\details{       
Additional arguments: 
\describe{
  \item{isToCopyEdges}{Logical value, whether to include edges to the copy (default = TRUE). }
  \item{isDefaultCopy}{Logical value, whether to duplicate the complete network or to copy only the original graph (default = TRUE). } 
  \item{nodes}{Optional: nodes to be duplicated <array of strings> (p.s. in this case, isDefaultCopy=TRUE). }
}
}

\value{
Duplicates graphs in RedeR app. 
}

\author{Mauro Castro}

\examples{
# Deprecated function...
}

\keyword{graph}
