% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\docType{methods}
\name{exitd,RedPort-method}
\alias{exitd,RedPort-method}
\alias{exitd}
\title{Exit the RedeR R-to-Java interface}
\usage{
\S4method{exitd}{RedPort}(obj)
}
\arguments{
\item{obj}{A \code{RedPort}-class object.}
}
\value{
Exit/close the RedeR application.
}
\description{
Close an active RedeR session.
}
\examples{

rdp <- RedPort()

\donttest{
calld(rdp)
exitd(rdp)
}

}
\seealso{
\code{\link{exitRedeR}}
}
\author{
Sysbiolab.
}
