\name{startRedeR}
\alias{startRedeR}
\alias{addGraphToRedeR}
\alias{resetRedeR}
\alias{getGraphFromRedeR}
\alias{relaxRedeR}
\alias{addLegendToRedeR}

\title{Warnings for RedeR transitional version.}

\usage{
startRedeR(...)
addGraphToRedeR(...)
resetRedeR(...)
getGraphFromRedeR(...)
relaxRedeR(...)
addLegendToRedeR(...)
}

\arguments{
\item{...}{not used.}
}

\value{
Warnings
}

\description{
These new functions are part of a new launch pipeline introduced in RedeR version 3 or later. However, a user might attempt to use them on an outdated version of RedeR. To mitigate this, a warning will be issued on the transitional version.
}

\examples{
# Return warnings.
}

