% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\docType{methods}
\name{ping,RedPort-method}
\alias{ping,RedPort-method}
\alias{ping}
\title{Ping RedeR app}
\usage{
\S4method{ping}{RedPort}(obj)
}
\arguments{
\item{obj}{A \code{RedPort}-class object.}
}
\value{
Ping test for RedeR app, either '1' (accessible) or '0'
(not accessible)
}
\description{
Test the R-to-Java interface of an active RedeR session.
}
\examples{

rdp <- RedPort('MyPort')

\donttest{
ping(rdp)
# [1] 0
calld(rdp)
ping(rdp)
# [1] 1
}

}
\seealso{
\code{\link{pingRedeR}}
}
\author{
Sysbiolab.
}
