% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\docType{data}
\name{RedeR-data1}
\alias{RedeR-data1}
\alias{hs.inter}
\title{Pre-processed igraph object for RedeR case studies.}
\format{
igraph
}
\source{
This package.
}
\usage{
data(hs.inter)
}
\value{
A pre-processed igraph object.
}
\description{
Preprocessed Human interactome extracted from the Human
Protein Reference Database (HPRD) in April 2011 <igraph object>
('name' attribute is mapped to ENTREZ ID).
}
\examples{
data(hs.inter)
}
\keyword{hs.inter}
