% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\name{rederInfo}
\alias{rederInfo}
\title{RedeR Attribute Information}
\usage{
rederInfo(what = c("Usage", "Value", "All"))
}
\arguments{
\item{what}{Character string specifying which attribute fields to return.}
}
\value{
Either display info summary or a  named list with three elements:
\describe{
  \item{Graph}{Data frame of graph-level attributes.}
  \item{Vertex}{Data frame of vertex-level attributes.}
  \item{Edge}{Data frame of edge-level attributes.}
}
}
\description{
Retrieves information on graph-, vertex-, and edge-level attributes recognized
by the RedeR application, including usage descriptions and/or value
specifications.
}
\examples{
## List usage information
rederInfo("Usage")

## List value information
rederInfo("Value")

## List both usage and value information
rederInfo("All")

}
\author{
Sysbiolab.
}
