% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\name{RedPort}
\alias{RedPort}
\title{Constructor of RedPort-class objects}
\usage{
RedPort(title = "default", host = "127.0.0.1", port = 9091, checkJava = FALSE)
}
\arguments{
\item{title}{A string naming the RedeR interface.}

\item{host}{Domain name of the remote computer that is running
the interface.}

\item{port}{An integer specifying the port on which the interface should
listen for incoming requests.}

\item{checkJava}{A logical value, whether to check the Java Runtime
Environment (JRE) installed on the system.}
}
\value{
An object of the RedPort class.
}
\description{
Constructor of the RedeR interface for remote procedure calls.
}
\examples{

# Initialize RedeR
library(RedeR)

rdp <- RedPort()

# Set global options used in internal methods
options(RedPort = RedPort())

}
\seealso{
\code{\link{startRedeR}}
}
\author{
Sysbiolab.
}
