% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedCallMethods.R
\docType{methods}
\name{resetd,RedPort-method}
\alias{resetd,RedPort-method}
\alias{resetd}
\title{Reset RedeR app}
\usage{
\S4method{resetd}{RedPort}(obj)
}
\arguments{
\item{obj}{A \code{RedPort}-class object.}
}
\value{
Reset plotting panel.
}
\description{
Reset an active RedeR session.
}
\examples{

rdp <- RedPort()

\donttest{
calld(rdp)
resetd(rdp)
}

}
\seealso{
\code{\link{resetRedeR}}
}
\author{
Sysbiolab.
}
