% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedWrappers.R
\name{resetRedeR}
\alias{resetRedeR}
\title{Reset RedeR app}
\usage{
resetRedeR()
}
\value{
Reset plotting panel.
}
\description{
Reset an active RedeR session.
}
\examples{
# Load RedeR
library(RedeR)

\donttest{
# Call 'start' and 'reset' methods
startRedeR()
resetRedeR()
}

}
\seealso{
\code{\link{startRedeR}}, \code{\link{addGraphToRedeR}}
}
\author{
Sysbiolab.
}
