% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{subg}
\alias{subg}
\title{Subgraph of a graph}
\usage{
subg(g, dat, refcol = 1, maincomp = TRUE, connected = TRUE, transdat = TRUE)
}
\arguments{
\item{g}{An 'igraph' object.}

\item{dat}{A data frame with node names and attributes to be mapped to 'g'.}

\item{refcol}{The reference column (node names) in the 'dat' object.}

\item{maincomp}{Logical value, whether to return only the main component
of the subgraph.}

\item{connected}{Logical value, whether to return only connected nodes.}

\item{transdat}{Logical value, whether to transfer node attributes
from the 'dat' object to the subgraph.}
}
\value{
An igraph object.
}
\description{
Creates a subgraph containing nodes specified from a data frame.
}
\examples{
# see 'nested subgraphs' section in RedeR's vignette:
# vignette("RedeR")

}
\seealso{
\code{\link[igraph:subgraph]{subgraph}}
}
