\name{gtoy.rm}
\alias{gtoy.rm}

\title{Random graphs and modules.}

\description{
A simple function to simulate random graphs with modular structures.
}

\usage{
gtoy.rm(m=5, nmax=30, nmin=5, p1=0.5, p2=0.03, p3=0.7, gg=NULL, 
nn=vcount(gg), noise.range=c(0.2, 0.6), plot=FALSE, fname=NULL)
}

\arguments{
  \item{m}{ Number of modules. }
  \item{nmax}{ The maximum number of vertices in each module. }  
  \item{nmin}{ The minimum number of vertices in each module. }  
  \item{p1}{ Probability for adding new vertices to a module. }  
  \item{p2}{ Probability for drawing an edge between modules. }
  \item{p3}{ Probability for drawing an edge within modules. }
  \item{gg}{ 
  An igraph object generated by the 'gtoy.rm' function.
  This option will add a random graph to the 'gg' object, following 
  Erdos-Renyi model (see \code{\link[igraph:erdos.renyi.game]{erdos.renyi.game}}).
  }
  \item{nn}{
  When 'gg' is provided, this option defines the number of new vertices 
  to add to the 'gg' object.
  }
  \item{noise.range}{
  When 'gg' is provided, this option sets the upper and lower
  bounds of a uniform random distribution assigned to the random graph 
  by the \code{\link[stats:runif]{runif}} function, which is applied with
  default settings.
  }
  \item{plot}{ 
  A logical value. This option generates a boxplot for the simulated signal 
  generated for each graph module.
  }
  \item{fname}{
  A string. The name of the pdf file in which the plot will be saved.
  }
}

\value{
Returns a igraph object. 
}

\author{Mauro Castro}

\examples{

#g<-gtoy.rm()

}

\keyword{graph}
