\name{RedeR-package}
\alias{RedeR-package}
\alias{RedeR}
\docType{package}
\title{
RedeR: bridging the gap between network analysis and visualization.
}
\description{
RedeR is an R-based package combined with a stand-alone Java application for interactive visualization and manipulation of modular structures, nested networks and multiple levels of hierarchical associations. The software takes advantage of R to run robust statistics, while the R-to-Java interface bridges the gap between network analysis and visualization.
}

\details{
\tabular{ll}{
Package: \tab RedeR \cr
Type: \tab Package\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}

\seealso{
\code{\link[RedeR:RedPort-class]{RedPort-class}}
}

\author{
Mauro Castro <mauro.a.castro@gmail.com>
}
\references{
Castro, MAA et al. \emph{RedeR: R/Bioconductor package for representing modular structures, nested networks and multiple levels of hierarchical associations}. Genome Biology 13(4):R29, 2012.
}
\keyword{ package }
