% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RedPort-class}
\alias{RedPort-class}
\title{RedPort: An S4 class for RedeR graphics}
\value{
An S4 class object.
}
\description{
RedPort: An S4 class for RedeR graphics
}
\section{Slots}{

\describe{
\item{\code{title}}{A string naming the XML-RPC port.}

\item{\code{uri}}{The uri to the XML-RPC server.}

\item{\code{host}}{Domain name of the machine that is running the XML-RPC server.}

\item{\code{port}}{An integer specifying the port on which the XML-RPC server
should listen.}
}}

\section{Constructor}{
 \code{\link{RedPort}}
}

\seealso{
\code{\link{addGraph}}, \code{\link{getGraph}}, \code{\link{relax}},
\code{\link{calld}}, @seealso \code{\link{resetd}}, \code{\link{exitd}},
\code{\link{ping}}, \code{\link{version}}.
}
\author{
Sysbiolab.
}
